@@ -7,10 +7,10 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Identity;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using WithoutIdentity.Data;
 using WithoutIdentity.Models;
 
@@ -28,6 +28,16 @@ namespace WithoutIdentity
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
+            string connString = Configuration.GetConnectionString("IdentityConnString");
+            Console.WriteLine(connString);
+
+            services.AddDbContext<ApplicationDataContext>(options =>
+                options.UseSqlServer(connString)
+            );
+            services.AddIdentity<ApplicationUser, IdentityRole<Guid>>()
+                   .AddEntityFrameworkStores<ApplicationDataContext>()
+                   .AddDefaultTokenProviders();
+
             services.Configure<CookiePolicyOptions>(options =>
             {
                 // This lambda determines whether user consent for non-essential cookies is needed for a given request.
@@ -35,14 +45,10 @@ namespace WithoutIdentity
                 options.MinimumSameSitePolicy = SameSiteMode.None;
             });
 
-            string connString = Configuration.GetConnectionString("IdentityConnString");
-            Console.WriteLine(connString);
 
-            services.AddDbContext<ApplicationDataContext>(options => options.UseSqlServer(connString));
 
-            services.AddIdentity<ApplicationUser, IdentityRole<Guid>>()
-                    .AddEntityFrameworkStores<ApplicationDataContext>()
-                    .AddDefaultTokenProviders();
+           
+
             services.Configure<IdentityOptions>(options =>
             {
                 options.Password.RequireDigit = false;
@@ -56,24 +62,28 @@ namespace WithoutIdentity
 
             services.ConfigureApplicationCookie(options =>
             {
-                options.CookieHttpOnly = true;
+                options.Cookie.HttpOnly = true;
                 options.ExpireTimeSpan = TimeSpan.FromMinutes(5);
                 options.LoginPath = "/Account/Login";
                 options.LogoutPath = "/Account/Logout";
                 options.AccessDeniedPath = "/Account/AccessDenied";
                 options.SlidingExpiration = true;
-  
+
 
             });
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+
+
+            services.AddControllersWithViews();
+            services.AddRazorPages();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
+                app.UseDatabaseErrorPage();
             }
             else
             {
@@ -81,19 +91,20 @@ namespace WithoutIdentity
                 // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
                 app.UseHsts();
             }
-
-            app.UseAuthentication();
-
             app.UseHttpsRedirection();
             app.UseStaticFiles();
-            app.UseCookiePolicy();
 
+            app.UseRouting();
 
-            app.UseMvc(routes =>
+            app.UseAuthentication();
+            app.UseAuthorization();
+
+            app.UseEndpoints(endpoints =>
             {
-                routes.MapRoute(
+                endpoints.MapControllerRoute(
                     name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
+                    pattern: "{controller=Home}/{action=Index}/{id?}");
+                endpoints.MapRazorPages();
             });
         }
     }
