@@ -1,18 +1,17 @@
-
-@model RegisterViewModel
-
+﻿@model LoginViewModel
 @{
-    ViewData["Title"] = "Registro";
+    ViewData["Title"] = "Login";
     Layout = "~/Views/Shared/_Layout.cshtml";
 }
 
+<h1>Login</h1>
 
 <h2>Registro</h2>
 
 <div class="row">
     <div class="col-md-4">
         <form asp-route-returnUrl="@ViewData["ReturnUrl"]" method="post">
-            <h4>Criar uma nova conta</h4>
+            <h4>Utilize uma conta para realizar o login</h4>
             <hr />
             <div class="text-danger" asp-validation-summary="All"></div>
 
@@ -28,14 +27,19 @@
                 <span asp-validation-for="Password" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <label asp-for="ConfirmPassword"></label>
-                <input asp-for="ConfirmPassword" type="password" value="" class="form-control" />
-                <span asp-validation-for="ConfirmPassword" class="text-danger"></span>
+                <div class="checkbox">
+                    <label asp-for="RememberMe">
+                        <input asp-for="RememberMe"/>
+                        @Html.DisplayNameFor(x=> x.RememberMe)
+                    </label>
+                </div>
             </div>
-            <button class="btn btn-default" type="submit">Criar</button>
+            <div class="form-group">
+                <button class="btn btn-default" type="submit">Login</button>
+            </div>
         </form>
     </div>
 </div>
-@section Scripts{ 
+@section Scripts{
     @await Html.PartialAsync("_ValidationScriptsPartial")
 }
