@@ -1,6 +1,6 @@
 ﻿@using WithoutIdentity
 @using WithoutIdentity.Models
-@Using Microsoft.AspNetCore.Identity
+@using Microsoft.AspNetCore.Identity
 @using WithoutIdentity.Models.AccountViewModel
+@using WithoutIdentity.Models.ManagerViewModel
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
-  
